move route handlers and helpers out of server module
